task: move task_complete ahead task function run
With IPC task, the function will free lock on IPC and allow new IPC
comes from HOST. But the state of ipc_task may be not udpated at this
moment, and the new comer IPC will be droped by task scheduler add.
Move the complete ahead to avoid this situation. May need more
attention when new user for task.

Signed-off-by: Pan Xiuli <xiuli.pan@linux.intel.com>